:recycle: create useHideOnBlur hook
