+++ client/utils/custom-hooks.js
@@ -13,3 +13,21 @@ export const useDidUpdate = (callback, deps) => {
     }
   }, deps);
 };
+
+export const useHideOnBlur = (hideOverlay) => {
+  const ref = useRef({});
+  const setRef = (r) => { ref.current = r; };
+
+  const handleClickOutside = ({ target }) => {
+    if (ref && ref.current && !ref.current.contains(target)) {
+      hideOverlay();
+    }
+  };
+
+  useEffect(() => {
+    document.addEventListener('mousedown', handleClickOutside);
+    return () => document.removeEventListener('mousedown', handleClickOutside);
+  }, [ref]);
+
+  return setRef;
+};
